Extract plugin code to a separate module
