+++ flow-plugins/flow-maven-plugin/src/test/resources/jar-with-frontend-resources.jar
