+++ flow-plugins/flow-maven-plugin/src/test/resources/jar-without-frontend-resources.jar
