+++ pom.xml
@@ -31,7 +31,7 @@
         <module>flow-tests</module>
         <module>flow-server-production-mode</module>
         <module>flow-component-demo-helpers</module>
-        <module>flow-maven-plugin</module>
+        <module>flow-plugins</module>
         <module>flow-test-generic</module>
         <module>flow-bom</module>
         <module>fusion-endpoint</module>
